+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Update Moodle plugin CI to include Moodle 5.2 and all supported PHP versions
 - Add sub-plugin PHPUnit test execution stage to Moodle plugin CI pipeline
 - Add support for receiving chunked uploads of quiz archives
+- Add an option to flatten the archive structure, placing all files directly in the root directory of the archive.
+- Allow exporting only the latest quiz attempt of each user in the generated archive
+- Add an option to include or exclude the quiz attempts metadata CSV file
 - Add support for question type [JACK](https://github.com/Wunderbyte-GmbH/moodle_qtype_jack) in quizes.
 - Include core sub-plugins in PHPUnit coverage reports
 - Adapt unit tests to Moodle upstream permission checks
@@ -12,6 +15,7 @@
 - Move sub-plugin directories to plugin root to align with Moodle core conventions
 - Restructure repository to meet Moodle coding style level 2 namespace suggestions
 - Improve Moodle "pluginfile image" inlining logic in report generators
+- Improve handling of duplicate files in archives
 
 
 ## Version 1.0.0 (2025112300)